Update demo 2
