--- Teams/04 Fundamentals of Microsoft Teams/Demos/02 - teams-bot1/teams-bot1.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27428.2015
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "teams-bot1", "teams-bot1\teams-bot1.csproj", "{2C9A9BD5-A8CD-4880-AE4E-555AF95FD0F0}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{2C9A9BD5-A8CD-4880-AE4E-555AF95FD0F0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2C9A9BD5-A8CD-4880-AE4E-555AF95FD0F0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2C9A9BD5-A8CD-4880-AE4E-555AF95FD0F0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2C9A9BD5-A8CD-4880-AE4E-555AF95FD0F0}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {E7B4D6F8-3432-4941-A85D-2FF9F35C1F1B}
-	EndGlobalSection
-EndGlobal
