+++ Teams/04 Fundamentals of Microsoft Teams/Demos/02-teams-bot-visual-studio/solution/teams-bot1.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.28010.2019
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "teams-bot1", "teams-bot1\teams-bot1.csproj", "{54738E8A-0B1E-4583-8443-917856BC5B46}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{54738E8A-0B1E-4583-8443-917856BC5B46}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{54738E8A-0B1E-4583-8443-917856BC5B46}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{54738E8A-0B1E-4583-8443-917856BC5B46}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{54738E8A-0B1E-4583-8443-917856BC5B46}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {82FD0A0C-7877-47CA-949A-D83E4569BBF8}
+	EndGlobalSection
+EndGlobal
