+++ Teams/04 Fundamentals of Microsoft Teams/Demos/02-teams-bot-visual-studio/solution/teams-bot1/App_Start/WebApiConfig.cs
@@ -1,0 +1,37 @@
+﻿using Newtonsoft.Json;
+using Newtonsoft.Json.Serialization;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web.Http;
+
+namespace teams_bot1
+{
+	public static class WebApiConfig
+	{
+		public static void Register(HttpConfiguration config)
+		{
+			// Json settings
+			config.Formatters.JsonFormatter.SerializerSettings.NullValueHandling = NullValueHandling.Ignore;
+			config.Formatters.JsonFormatter.SerializerSettings.ContractResolver = new CamelCasePropertyNamesContractResolver();
+			config.Formatters.JsonFormatter.SerializerSettings.Formatting = Formatting.Indented;
+			JsonConvert.DefaultSettings = () => new JsonSerializerSettings()
+			{
+				ContractResolver = new CamelCasePropertyNamesContractResolver(),
+				Formatting = Newtonsoft.Json.Formatting.Indented,
+				NullValueHandling = NullValueHandling.Ignore,
+			};
+
+			// Web API configuration and services
+
+			// Web API routes
+			config.MapHttpAttributeRoutes();
+
+			config.Routes.MapHttpRoute(
+					name: "DefaultApi",
+					routeTemplate: "api/{controller}/{id}",
+					defaults: new { id = RouteParameter.Optional }
+			);
+		}
+	}
+}
