+++ Teams/04 Fundamentals of Microsoft Teams/Demos/02-teams-bot-visual-studio/solution/teams-bot1/Dialogs/RootDialog.cs
@@ -1,0 +1,31 @@
+﻿using System;
+using System.Threading.Tasks;
+using Microsoft.Bot.Builder.Dialogs;
+using Microsoft.Bot.Connector;
+
+namespace teams_bot1.Dialogs
+{
+	[Serializable]
+	public class RootDialog : IDialog<object>
+	{
+		public Task StartAsync(IDialogContext context)
+		{
+			context.Wait(MessageReceivedAsync);
+
+			return Task.CompletedTask;
+		}
+
+		private async Task MessageReceivedAsync(IDialogContext context, IAwaitable<object> result)
+		{
+			var activity = await result as Activity;
+
+			// Calculate something for us to return
+			int length = (activity.Text ?? string.Empty).Length;
+
+			// Return our reply to the user
+			await context.PostAsync($"You sent {activity.Text} which was {length} characters");
+
+			context.Wait(MessageReceivedAsync);
+		}
+	}
+}
