+++ Teams/04 Fundamentals of Microsoft Teams/Demos/02-teams-bot-visual-studio/solution/teams-bot1/Web.config
@@ -10,7 +10,7 @@
   </configSections>
   <appSettings>
     <!-- update these with your BotId, Microsoft App Id and your Microsoft App Password-->
-    <add key="BotId" value="YourBotId" />
+    <add key="BotId" value="" />
     <add key="MicrosoftAppId" value="" />
     <add key="MicrosoftAppPassword" value="" />
   </appSettings>
@@ -34,7 +34,6 @@
         <add value="default.htm" />
       </files>
     </defaultDocument>
-
     <handlers>
       <remove name="ExtensionlessUrlHandler-Integrated-4.0" />
       <remove name="OPTIONSVerbHandler" />
@@ -62,7 +61,7 @@
       </dependentAssembly>
       <dependentAssembly>
         <assemblyIdentity name="Newtonsoft.Json" publicKeyToken="30ad4fe6b2a6aeed" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-8.0.0.0" newVersion="8.0.0.0" />
+        <bindingRedirect oldVersion="0.0.0.0-10.0.0.0" newVersion="10.0.0.0" />
       </dependentAssembly>
       <dependentAssembly>
         <assemblyIdentity name="System.Net.Http.Primitives" publicKeyToken="b03f5f7f11d50a3a" culture="neutral" />
