+++ Teams/04 Fundamentals of Microsoft Teams/Demos/02-teams-bot-visual-studio/solution/teams-bot1/default.htm
@@ -1,0 +1,27 @@
+﻿<!DOCTYPE html>
+<html>
+<head>
+    <title></title>
+    <meta charset="utf-8" />
+</head>
+<body style="font-family:'Segoe UI'">
+    <h1>teams_bot1</h1>
+    <p>This is a simple echo bot sample using Microsoft Bot Framework.</p>
+    <p>Visit <a href="https://www.botframework.com/">Bot Framework</a> to register your bot. When you register it, remember to set your bot's endpoint to <pre>https://<i>your_bots_hostname</i>/api/messages</pre></p>
+    <p><a href="https://azuredeploy.net"><img title="" alt="Deploy to Azure" src="https://azuredeploy.net/deploybutton.png"></a></p>
+
+    <h3 id="prerequisites">Prerequisites</h3>
+    <p>
+        The minimum prerequisites to run this sample are: <br>
+        <ul>
+            <li>The latest update of <b>Visual Studio 2015, 2017</b>. You can download the community version <a href="http://www.visualstudio.com">here</a> for free.</li>
+            <li>
+                The <a href="file:///C:\Users\%USERPROFILE%\AppData\Local\botframework\app-3.5.29\botframework-emulator.exe">Bot Framework Emulator</a>. To install the Bot Framework Emulator, download it from <a href="https://emulator.botframework.com/">here</a>.
+                Once the emulator has been installed you can debug your sample by folowing the steps provided in <a href="https://docs.microsoft.com/en-us/bot-framework/debug-bots-emulator">this</a> document.
+                Please refer to <a href="https://github.com/microsoft/botframework-emulator/wiki/Getting-Started">this documentation article</a> to know more about the Bot Framework Emulator.
+            </li>
+        </ul>
+    </p>
+    <p>For more information on how to build bots you can visit the official Bot Framework documentation <a href="https://docs.microsoft.com/en-us/bot-framework/">here</a></p>
+</body>
+</html>
