+++ Teams/04 Fundamentals of Microsoft Teams/Demos/02-teams-bot-visual-studio/solution/teams-bot1/teams_bot1.bot
@@ -1,0 +1,14 @@
+{
+  "name": "teams_bot1",
+  "secretKey": "",
+  "services": [
+    {
+      "appId": "",
+      "id": "http://localhost:3978/api/messages",
+      "type": "endpoint",
+      "appPassword": "",
+      "endpoint": "http://localhost:3978/api/messages",
+      "name": "teams_bot1"
+    }
+  ]
+}
